Merge permission for modified Output into single permission
Based on feedback from Eric Shultz, I've merged the Javascript/CSS permission
to work the same way as the HTML permission.

The goal is to give permission for downstream to incorporate unmodified
HTML/CSS/Javascript with their own works, but assure that they don't copy
parts of the otherwise AGPLv3'd codebase into that Output.