Use "covered work", the defined term from AGPLv3 where possible.
AGPLv3 defines the term "covered work" already, which becomes the core phrase
of strong copyleft throughout the existing License.

Using this term allows for various simplifications to the permission
statement.

Furthermore, there is no reason that the licensor can (or really, should try
to) grant or copyright permissions for works that aren't covered works.

Pam Chestek originally gave me this idea by making her change to §2¶2,
pointing out that "works" was problematic there.

Finally, the use of the word "file" and "files" was already problematic.
Most of the CSS/Javascript/HTML might not be in "files" of its own -- it may
for example be inside print statements strewn throughout the covered work.
Referring to them as "files" gave the wrong impression to start, something
Eric Schultz had raised earlier in drafting.