Defined term "Output" should be "Covered Output" instead.
The word "output" (undefined) is used quite a bit in the body of AGPLv3.
Since an additional permission is read as part of the license itself, I don't
think we should define the term (notwithstanding the apparent case
sensitivity) -- particularly since some defined terms in AGPLv3 are not
capitalized.

Furthermore, this reinforces that we're primarily concerned about Output that
is a derived/derivative/combined work with a covered work under the License.