Ditch the idea of "Package"; makes it all simpler.
The idea of being able to apply this to part of a repository was nice, but it
complicates drafting by requiring us to define a term "Package" that is
somehow a subset of the "Program" and "unmodified Program", which are terms
already used in AGPLv3.

Perhaps later someone can find a way to extend this exception to work that
way, but not today.