Parity with two phrases already used in AGPLv3's main body.
AGPLv3§2 uses the phrase "output from running a covered work", so we use the
same phrase here.  §6(e) already uses the phrase "general public", and of
course "propagate" is a defined term from AGPLv3§0.

This reworked definition of "Covered Output" better normalizes its definition
to fit the standard terminology of AGPLv3, and also removes the 'deployed on
a website'.  We actually don't want to create the idea that deploying on the
website is the only way to be eligible for the additional permission.  The
key issue is to assure that any items that receive the additional permission
are distributed to the general public, as we want "one last step" of Affero
requirement for the Covered Output, which only once available to the public
can then be combined with other software more liberally.